refactor: product.tsx
